Candle value type fix

+++ client-api/src/main/java/com/artiow/moex/api/model/Candle.java
@@ -3,18 +3,19 @@ package com.artiow.moex.api.model;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
 @Setter
 public class Candle {
 
-    private Double open;
-    private Double close;
-    private Double high;
-    private Double low;
-    private Double value;
-    private Double volume;
+    private BigDecimal open;
+    private BigDecimal close;
+    private BigDecimal high;
+    private BigDecimal low;
+    private BigDecimal value;
+    private BigDecimal volume;
     private LocalDateTime begin;
     private LocalDateTime end;
 }
